Don't look for limactl on the PATH
And don't use LIMA_HOME from the process environment.

The embedded version of Lima and the location of the VM are supposed
to be completely separate from the user's installation and should
not interfere with each other. rdctl is only working with the
rancher desktop version of Lima and instance 0 in the expected location.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>