Relay error messages from the helper back to the user.
For all the functions, output should be empty for two of them,
and valid JSON for the others, so we can make that check and
return the output with status 200 if it passes, 400 if not.

Revise the e2e test to JSON.stringify the payloads, and then
passwords can contain non-alphanumeric characters.

Signed-off-by: Eric Promislow <epromislow@suse.com>