No need to pass `--tty --stdin` to `k exec` when not doing anything interactive.
Sometimes the CI fails the call to `kubectl exec NGINX-POD -- curl IP:PORT` with
the error message:

    stdout:
    stderr: Unable to use a TTY - input is not a terminal or the right kind of file
    Error from server: error dialing backend: EOF

There's no need to pass `--tty --stdin` to the call to `exec`.

Also filter out other pods for when this test is run locally on an existing
k8s backend, with other pods in the default namespace.

Signed-off-by: Eric Promislow <epromislow@suse.com>