Add a Kubernetes watcher.
This component connects to a Kubernetes cluster using a provided kubeconfig
file (defaulting to the k3s one), and watches for services.  This is used
to handle newer versions of Kubernetes which has stopped listening on the
local interface for NodePort services.

This is the initial proof of concept; much needs to change for actual use.

Signed-off-by: Mark Yen <mark.yen@suse.com>