Enable Kube NodePort forwarding via RD-guest-agent
In newer versions of Kubernetes, the kubelet will no longer listen on the
local port when setting up a NodePort service (which is actually handled
via iptables).  However, our platforms (WSL & lima) actually rely on that
listening local port to determine when they need to do port forwarding from
the host.  So we watch Kubernetes services and manually listen on the
corresponding port to emulate the old behaviour (and repairing the port
forwarding).

This introduces the rancher-desktop-guestagent service for lima; it
previously already existed on WSL for containerd-based port forwarding
(which is handled on lima via lima-guest-agent).

Signed-off-by: Mark Yen <mark.yen@suse.com>