Backend: use KubernetesBackend where needed.
Use the `kube` property from VMBackend where necessary, instead of using
the VMBackend directly as a KubernetesBackend; this means we can later not
assume it's a KubernetesBackend.  Note that this means some variables are
badly named for now; this will be cleaned up in a follow-up commit.

KubernetesBackend.isServiceReady was never used.

Signed-off-by: Mark Yen <mark.yen@suse.com>