WSL: Catch wsl.exe responding with usage error.
On Windows 11 (and possibly recent Windows 10), `wsl.exe` exists but is a
stub that is just used for installing the main package (which will
effectively replace `wsl.exe`).  When we run it initially, check for usage
messages in addition to checking for the exception being ENOENT to catch
when it's not installed.

Signed-off-by: Mark Yen <mark.yen@suse.com>