Diagnostics: implement one diagnostic.
This implements the connected-to-internet diagnostic.  Note that some of
the diagnostics structures have to be changed to make it possible to
actually run diagnostics.

The YAML listing is removed, as we need actual code to implement each
diagnostic (so it provided no actual value).

Actually providing fixes is not yet implemented.  The tests which
previously had fixes are commented out for now.

Signed-off-by: Mark Yen <mark.yen@suse.com>