Lima: Fix kube downgrade data deletion.
We accidentally dropped reading the existing Kubernetes version, and broke
downgrade detection.  Restore that (now by putting the file inside the VM
instead, to avoid issues with Lima warning about unknown configuration).

This copies some ideas from the WSL implementation (including the path of
the file where the version is stored in the VM).

Signed-off-by: Mark Yen <mark.yen@suse.com>