Try harder to kill the app when doing a manual factory reset in bats
When `rdctl shutdown` doesn't actually kill the app, then we need to
try harder ourselves before deleting the VM. Otherwise we can end up
with a running app that is missing the VM and throws curious errors.

Also set default memory size explicitly, as RD doesn't seem to apply
the default value if the setting is missing.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>