Lima: Fix generating initial sudoers file.
When we run `limactl sudoers`, it actually checks for the given executables
to exist before accepting them.  This conflicts with us wanting those
executables to be root-owned, and wanting a singular sudo prompt (so we
must generate the sudoers file before the files are in place).

As a workaround, initially create the networking configuration with the
executables from the application directory, and do a string replace on the
generated sudoers file to use the root-owned version instead.

Signed-off-by: Mark Yen <mark.yen@suse.com>