Run in debug mode when env var RD_DEBUG_ENABLED is set (to a non-empty string).
Now the `settings.debug` value and `RD_DEBUG_ENABLED` are treated independently,
where if `settings.debug` is true, we always run in debug mode. If it's false
but `RD_DEBUG_ENABLED` is set, we run in debug mode, but don't change the persisted
setting.

The main wrinkle is in how the debug checkbox is handled in the Troubleshooting UI:
If the RD_DEBUG_ENABLED flag is set, the button is checked but disabled.
If it isn't set, the button behaves as usual.

Signed-off-by: Eric Promislow <epromislow@suse.com>