Fix checkKubernetesVersion logic
The validator used to always return `true` (meaning the settings have
changed) when Kubernetes was disabled. This seems wrong.

The new code continues to check if the currentValue matches the
desiredVersion and only returns `true` when they are different.

The only thing that is different with Kubernetes being disabled is that
the empty string is a valid desired version.

Before this change we would get this illogical result when Kubernetes
was disabled:

```
$ rdctl api settings | rdctl api -X PUT --input - settings
settings updated; no restart required
```

After this change it produces the correct result, regardless whether
Kubernetes is disabled or not:

```
$ rdctl api settings | rdctl api -X PUT --input - settings
no changes necessary
```

Signed-off-by: Jan Dubois <jan.dubois@suse.com>