Lima/kube: don't use tail to copy logs.
On macOS Ventura (13), we are seeing hangs once Kubernetes starts up.  It
appears to have to do with the process tailing the k3s logs to copy it to
the host.

Rather than using an explicit process for that, just mount the logs
directory inside the VM (via sshfs built in to lima) and write the logs
there directly.  This appears to fix the issue.

Signed-off-by: Mark Yen <mark.yen@suse.com>