WiX: Detect when WSL has been installed from the Windows store.
If we've already been installed from the Window store, then we don't need
to install again.  Note that this currently doesn't affect the installer
flow (it still forces admin install, and requests a reboot at the end) even
though it's not necessary.  It is currently not possible to detect from the
MSI when this is the case; we'll probably need a custom action for that.

(While there's support for JScript-based custom actions, the advice from
the main WiX developer seems to be to avoid using it.
https://web.archive.org/web/20160217123733/http://blogs.msdn.com/b/robmen/archive/2004/05/20/136530.aspx )

Signed-off-by: Mark Yen <mark.yen@suse.com>