WiX: Assume WSL is installed if RD was previously installed.
Windows Installer refuses to locate a per-machine install of the WSL kernel
if we're installing per-user (when installing passively; when installing
with normal UI, we're flagged as per-machine when we get to that point and
only change to per-user afterwards).  This means we mistakenly think we
need to elevate to install WSL.

If we detect an upgrade of Rancher Desktop (either by having a previous
install of the same upgrade code, or by finding the NSIS uninstall string),
treat that as WSL was already installed.  This works around the
aforementioned issue (by relaxing the WSL requirement).

Signed-off-by: Mark Yen <mark.yen@suse.com>