WiX: Don't override MSIINSTALLPERUSER in the server.
MSI runs with UI first, and then reruns itself as the server.  After the
user has selected the configuration in the UI, all configuration is passed
to the server and the installer is run essentially as a silent install.
In that second run, we should not override MSIINSTALLPER user coming from
the first run.

To do so, we:
- Switch the <SetProperty> to Sequence="first"
- As a safeguard, only do so if either WSLKERNELINSTALLED or
  MSIINSTALLPERUSER is unset.  This ensures we do not regress
  bff47ac55f83781bc16a9e01482a1c45cc8b15c3.

Signed-off-by: Mark Yen <mark.yen@suse.com>