WiX: Fix upgrade install of per-user installs.
When we run the installer with full UI (i.e. the default mode), we are
initially set up as if we're going to do a per-machine installation, and
then this gets changed as the user steps through the dialogs.  This means
that when it initially runs the FindRelatedProducts step (looking for old
installations to upgrade from), this happens in the per-machine context,
and will skip over per-user installations.  If the user chooses a per-user
installation, this eventually leads to us not uninstalling the previous
installation before installing a major upgrade (in our case, this means any
other build).

Fix this by manually running FindRelatedProducts from the UI, once the
final confirmation before installation happens, after we have fixed up the
ALLUSERS / MSIINSTALLPERUSER properties to be a per-user installation.  At
that point the engine will pick up the previous per-user install and
therefore uninstall the previous installation before installing the new
build.

This commit also moves setting the per-machine/per-user properties at the
same place (after the user confirms installation) rather than where it was
previously (right after the scope selection dialog).  This is to fix an
issue where if the user chooses per-machine, move forward, then back again,
the UI gets confused (because the properties no longer exactly match one of
the expected values) and loses the selection.  We also now explicitly add
ordering to all the steps we do at install confirmation; it was previously
using implicit ordering (based on the position in the XML file).

Signed-off-by: Mark Yen <mark.yen@suse.com>