rdctl: Correctly kill Rancher Desktop on darwin.
On darwin, pkill by default does not kill ancestors of pkill.  However, we
spawn it from rdctl, which in turn was spawned from the Rancher Desktop
main process.  This means that we would fail to kill Rancher Desktop.  Fix
this by passing in the -a flag (only on darwin) to make it consider
ancestors.  Note that this is not an issue for qemu, since that isn't an
ancestor of pkill.

Signed-off-by: Mark Yen <mark.yen@suse.com>