rdctl: factory-reset: Manually handle killing processes.
When invoked from the GUI, rdctl is a child process of the main Rancher
Desktop process.  This means that invoking `taskkill ... /T` will also
terminate rdctl, which aborts the factory reset.

Instead, manually iterate through processes and terminating the ones that
are part of the application (that is, has an executable image that lives
within the application directory).  This is close enough to what we want to
achieve (and correctly kills vtunnel and host-resolver).

Also includes fixes to correctly bubble errors up.

Signed-off-by: Mark Yen <mark.yen@suse.com>