Get checksum using endsWith rather than includes
Recently, docker/buildx changed the way they do release assets:
for each executable, there are now two other assets that bear the
same name, but with extensions added on. This commit modifies
findChecksum so that the lines corresponding to the latter two
are filtered out, which allows it to work properly.

Signed-off-by: Adam Pickering <adam.pickering@suse.com>