nerdctl-stub: add tests for `nerdctl cp` handling.
Also change how the handling works to rely on the fact that we have exactly
two positional arguments, one of which refers to the container (and the
other refers to a host path or stdin/stdout).  This ensures we can better
detect which argument should refer to a host path (and only munge that one
as necessary).

Signed-off-by: Mark Yen <mark.yen@suse.com>