Actions: Rework how to detect if a secret exists.
Turns out that while we _can_ refer to secrets directly in the output,
they are missing values so things don't actually work.  We still need to
do the calculation in a step instead.

At least we _always_ had to have at least one step, so it's not actually
much slower.

Fixes: 6088f4a78f25b73efd3326aa88620b6bf7c18a7a
Signed-off-by: Mark Yen <mark.yen@suse.com>