Make some rdctl set/start options platform-specific.
If an option has an `x-rd-platforms` field, generate Go code
for that option only on the specified platforms.

On platforms that don't support the option, it's still invocable,
but doesn't appear in help, and an attempt to use it gives an
error message that it isn't available on the current platform.

Note that linux on WSL has the same limitations as native Windows.

Signed-off-by: Eric Promislow <epromislow@suse.com>

Process options that aren't meant for the current platform

and give an error message when the user tries to run one.
Note that platform-inappropriate options are "hidden"
and don't show up in help.

Signed-off-by: Eric Promislow <epromislow@suse.com>