Use "Open window" instead of "Show window"
"Open" is the proper opposite to "Close"; "Show" would imply that the
window has been hidden or minimized, but that is not the case: the
window has been destroyed and the app may not even be displayed in
the dock or taskbar anymore.

We also use "File | Close Window" in the app menu.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>