mainEvents: drop deprecated emit/on
Drop the generic versions of emit() and on(), which are confusing because
it's not always clear why they are deprecated.  Instead, implement the
MainEvents interface by forwarding to the methods from EventEmitter
manually.  This should reduce confusion about why a call isn't working
(since TypeScript can better guess which one is expected and emit better
errors about any arguments that are incorrect).

Signed-off-by: Mark Yen <mark.yen@suse.com>