mainEvents: support more methods from EventEmitter
We previous didn't do any type checking on these as they were not declared
on the interface and were just happened to work (because it was on
EventEmitter).  We now add checking for them.

This somehow lead to changes (around credentials) that are now needed due
to the methods being exposed...

Signed-off-by: Mark Yen <mark.yen@suse.com>