background: fix signature for writeSettings()
It appears that the JetBrains TypeScript checker thinks there is an issue
with how we call writeSettings() (and produces an error that doesn't make
much sense).  However, as we never modify the input here, we can work
around the issue by making the input also read only.

(The details are somewhat vague, as I could not reproduce the issue due to
lack of JetBrains licenses.)

Signed-off-by: Mark Yen <mark.yen@suse.com>