Redo Windows registry-based deployment profile reading.
Reading the registry is a bit strange because sub-object names (keys)
live in a different namespace from sub-object named values.

But we can lump both sets of names together, and then decide, based on
the type of the value of a given name in the schema, whether it's
a value in the registry to harvest, or a sub-object key to do further
matching on.

Yeah, it's a bit hard to explain. But this is why there's only one recursive
call to `readRegistryUsingSchema`.

Signed-off-by: Eric Promislow <epromislow@suse.com>