Add a one-second pause so containerd can get to work successfully.
I was able to run the code in this test manually, but it always fails
when it tries to run `nerdctl build --tag rd/extension/ui --build-arg variant=ui dataDir`

The error message showed that containerd.sock wasn't available.

Obviously when I ran the code interactively I gave the backend time to create the socket file.

If I just run `rdctl shell test -S .../containerd.sock` it fails to run `nerdctl build`,
but if I have it wait a second it works.

Signed-off-by: Eric Promislow <epromislow@suse.com>