RDX: Generate compose file for image-type extensions
We were previously manually running containers (instead of using a compose
file); this seems to cause issues on restart / re-install because the
container would be already running when we try to recreate it.  By using a
compose file instead, we can make the container engine do whatever is
required to reach the desired state instead.

Signed-off-by: Mark Yen <mark.yen@suse.com>