backend: nerdctl client: limit files when copying out of VM
Some VM images may contain symlinks to the root of the disk (e.g., /var/run
pointing to /run is popular).  This can create issues when we try to make a
tarball of the image, as it runs into various dev/udev/proc/etc. files.
Limit the archive to --one-file-system to avoid this.

Signed-off-by: Mark Yen <mark.yen@suse.com>