RDX: Add golang helper to forward HTTP-over-TCP to unix socket
Extensions can specify an (undocumented) socket to expose to the front end,
which will be used when the front end makes a ddClient.extension.vm.service
call that is relative (i.e. doesn't include scheme and host).  This proxy
handles the last leg, where some TCP traffic is passed to the Unix socket.
It is hard coded to listen on port 80 (which will be mapped to an arbitrary
port via compose).

Signed-off-by: Mark Yen <mark.yen@suse.com>