RDX: Inject backend proxying container
This injects a sidecar container when we need to expose a port from the
backend to the frontend, using the image we built in the previous commits.
This means that we have a port we can look up (using `docker compose port`)
that we can then connect to from the frontend.

Signed-off-by: Mark Yen <mark.yen@suse.com>