RDX: Don't send events to dead frames
If the user navigates away from the extension (causing the BrowserView to
be destroyed), we can't send things to that frame anymore.  But we might
still get results back from the process in that case.  Just ignore it in
that case.

We're not terminating the process here as it may be a long-running process
(e.g. `helm install ...`) that the user was just bored of monitoring; it's
likely that we should let it remain.

Signed-off-by: Mark Yen <mark.yen@suse.com>