wsl-exec: ensure we use the pid of /sbin/init
When _not_ using network namespaces, the pid written (from wsl-init) is the
pid of the unshare process that is the parent of /sbin/init.  While that
seems to mostly work (/sbin/init is pid 1, and we can't see the rest of the
pids), it somehow breaks rc-service (it finds the services but treats them
as unsupervised).  However, if we manually enter the pid namespace of
/sbin/init it works fine:

# /usr/bin/nsenter -t 1 /sbin/rc-service buildkitd status
 * status: unsupervised
# /usr/bin/nsenter -p -t 1 /sbin/rc-service buildkitd status
 * status: started

Fix this by switching to using /sbin/init (rather than unshare) in the
initial esenter call, even though it looks the same from a glance.

Signed-off-by: Mark Yen <mark.yen@suse.com>