RDX: Strip quotes around command lines
It appears that some extensions pass in unncessary quotes around some
parameters that would have caused issues in a shell.  It's likely that
Docker Desktop combined the arguments into a command line and executed it
with /bin/sh or similar.  As we never serialize the command line like that,
try to be compatible with the easy cases by stripping quotes.

Signed-off-by: Mark Yen <mark.yen@suse.com>