container client: add method to get tags of an image
This adds a method on the container engine clients to return the
available tags for a repository.  This includes any tags available
remotely (in the corresponding registry), as well as locally (already
downloaded / built).  We ignore any errors fetching tags, as the
registry is expected to return errors about repositories that were never
pushed.

Signed-off-by: Mark Yen <mark.yen@suse.com>