Don't `pkill -f rancher-desktop`
This breaks BATS when the local repo clone is called `rancher-desktop`,
which is not a completely unreasonable choice. :D

Use the full path to the resources directory instead to avoid killing
parts of the BATS framework. Also need to resolve symlinks to match
output of `ps`.

The `pkill` logic should also only be executed when we are running
with `npm run dev` because `rdctl factory-reset` doesn't know how to
shut down those instances. We should not have a fallback for the
normal use case; if `rdctl factory-reset` can fail, then that needs to
be fixed instead.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>