Logging: Add helper for printing message-and-error
The new helper, Logging#debugE, takes a message and an exception.  If we're
running in development mode, or if we're running under test (E2E or BATS),
print the message followed by the exception with stack trace.  Otherwise,
print a single line with the message and the exception, but without the
stack trace so that it's less likely to become a red herring when the real
issue is elsewhere.

Signed-off-by: Mark Yen <mark.yen@suse.com>