Add needs_port() function to BATS
Use it to request binding to a privileged port. It will succeed if the port
is greater equal to net.ipv4.ip_unprivileged_port_start. If it isn't, then
try to change the starting port using sudo. If that fails (e.g. because sudo
requires a password), then skip the test.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>