Add the `--output reg...` option to `rdctl list-settings`.
This is used to convert current settings into Windows registry file format.

* Generate a full options struct in golang representing all possible
  preference fields, including:
  * arrays
  * maps
  * fields used only for other platforms.

This is needed for this branch in order to convert JSON documents
containing the above constructs into reg-files, even if those features
aren't supported on the current platform.

This does not add support for specifying array or map values to
`rdctl set` or `rdctl start`, but is a pre-rec for those features.

* Reg files need to have scalar values appearing before nested structs.

  Otherwise, if scalar values appear after a nested struct is written out,
  it looks like the scalar belongs to the last nested struct, not the one
  it actually belongs to.

Signed-off-by: Eric Promislow <epromislow@suse.com>