Fix the redirection to localhost within the WSL VM
When spawning a container that use the port `80` with for instance the
following command `docker run -d -p  8000:80 nginx`. In your WSL VM if
you run a `curl` command on the port 8000 the translation to port 80
will redirect the traffic to your configured proxy.
This fix this issue by adding an iptables rule to not redirect any of
the traffic to `127.0.0.1`.

Signed-off-by: Thomas Perale <thomas.perale@mind.be>