Add global_teardown() function to BATS
That way even tests that override teardown_file() can still call it
at the end of their custom teardown function.

Ideally we would be using teardown_suite(), but that function must be
defined in a setup_suite.bash file in the same directory as the first
*.bats file passed to the bats command. That doesn't work for us, so you
would have to specify `--setup-suite-file tests/helpers/setup-suite.bash`
which is too cumbersome unless/until we use wrapper scripts to run tests.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>