Add BATS test filename in front of each file's output
This is needed for running bats in CI, in which case it will produce
TAP output covering all input files, so it is no longer easy to connect
the test output back to the test file.

This info is redundant when running bats from a terminal, but I haven't
found a way to tell the difference from inside the test helpers.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>