Don't throw an error that Rosetta is still selected when switching back to QEMU
It should be silently ignored. It is only an error when Rosetta was not
selected while using VZ; you can't enable it, and switch to QEMU in one
step, as it makes no sense.

Unfortunately we can't do the same thing for VIRTIOFS; the user **must**
select an alternative mounting mechanism before switching back to QEMU
as there are multiple choices.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>