Get rid of PathManagementStrategy.NotSet
It was introduced to help users upgrading from 1.3 to 1.4 (I think),
which added the path management strategy setting as a replacement for
putting symlinks into /usr/local/bin. Instead of the full first-run
dialog we would display a simpler version that only asks about the
value for the path management strategy.

This dialog has since been removed. If somebody will upgrade from 1.3
or earlier to 1.9 or later, then we'll default the strategy to
RcFiles. This can still be changed in the preferences dialog later.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>