E2E: Remove helm & kubectl tests
We already have a helm test in BATS (which installs Rancher Manager).  This
test is redundant, and E2E tests have tighter timing requirements.

The kubectl test was more useful when we were less stable; at this point it
is rather stable, so it's not bringing us any value (over the helm tests).

Signed-off-by: Mark Yen <mark.yen@suse.com>