Ignore case for the allowed specified values for `--reg-hive` and `--section`.
Noted that the allowed values for `--output` are case-*sensitive*.

The difference is that it's helpful if the user can use the same case that the
registry uses, namely all-caps for `--reg-hive` (e.g. HKLM_LOCAL_MACHINE`),
and capitalized for `--section` (like 'Defaults').  `reg` and `json` are
names of output-formats that we are proposing.

Also I don't convert the names to all-lower-case until I know we're using
them, so I have the original case if I need to put the entered string
in an error message.

Signed-off-by: Eric Promislow <epromislow@suse.com>