try() is not a variant of run()
It will fail with a non-zero exit code when not successful, so there is
no need to `assert_success` after calling `try` without `run`. And while
it does capture the command output in $output, that is an implementation
detail that tests should not rely on; they have to call `run try` to set
$output and $status.

Exiting with non-zero status was part of 431fa12 and outputting the
command output back to stdout was implemented in d678dba.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>