Split postinstall.ts into two parts, and provide a simple spawner with no other dependencies.
The problem is that some of the postinstall code was loading the logger,
which was loading `utils/paths.ts`, which depended on finding a working `rdctl`.

The first part of the solution is to build `rdctl` earlier and install it in a location where
the `paths.ts` code expects to find it.

The second part is to break up `postinstall.ts` into two parts, and use the simple-spawner
to execute each part, so we don't hit the logger or the paths.ts module until we're ready to.

Signed-off-by: Eric Promislow <epromislow@suse.com>