Make windows factory-reset sudoless
The script utilizes a WSL call to trigger a command-line
call inside the rancher-desktop distribution. This allows
executing iptables flush and deletion commands within the
actual rancher-desktop distribution. Previously, these
commands were run in the test (ubuntu) distribution, which
used a different iptables configuration than the one employed
by the rancher desktop.

Furthermore, the script addresses the issue with the xargs -l option.
This option is not available in busybox

Signed-off-by: Nino Kodabande <nkodabande@suse.com>