Retry updating the config.
Sometimes attempts to update the config file fail with an EBUSY error.
Most likely the test code has closed the file but the OS hasn't released
the write-lock on it, so retry.

I didn't see this failure on every run, maybe about 1/3 of the time.
Now I see that when it does fail, it only needs to retry once to succeed.

Signed-off-by: Eric Promislow <epromislow@suse.com>