First set of e2e deployment profile tests
These tests verify whether we see or don't see the first-run dialog box,
based on whether the settings and profiles exist when Rancher Desktop starts up.

In many cases, these conditions need to be set up manually,
so these tests aren't intended to be run as part of the regular e2e suite.

Signed-off-by: Eric Promislow <epromislow@suse.com>