Fix both the bats test for checking deleted directories, and the deleter.
On the rdctl side:
If the <apphome> directory doesn't contain a `snapshots` subdirectory,
delete it. Otherwise delete all of its contents except `snapshots`.

On the testing side:
Specify exactly which members of <apphome>/ we expect to see or not see.
Ignore the existence of a  `snapshots` subdirectory -- that all needs
to be done in a separate test that doesn't use this before/after assert/refute
dichotomy.

Signed-off-by: Eric Promislow <epromislow@suse.com>