The app doesn't always stop cleanly, but it does stop.
If `lima.start()` is interrupted by trying to select an
invalid k8s version, it's possible for `lima.stop()` to
throw an exception (like if it tries to stop a service
like buildkitd, which was never started).

So just look for the "child" process to have ended,
and then check for locked-field error messages.

Signed-off-by: Eric Promislow <epromislow@suse.com>