Move everything from the roaming to local profile.
Follow-up from moving the location of settings.json on Windows:

If it's not in the new location, check to see if it's in the
old location, and copy it over to the new location if it is.

Keep doing this as long as people run version 1.10 and then migrate
forward. We copy the file instead of moving it so users can fallback
to version 1.10 and have their old settings if they need to.

A factory-reset will delete settings.json from both the old and new
locations.

Also (post-review changes):
- turn comments in bats tests into separate test headers
- AppData should always be called RoamingAppData, but this PR might not get all of them

Signed-off-by: Eric Promislow <epromislow@suse.com>