Switch to using .tar format for snapshots instead of .vhdx format
Exporting/importing via .tar is slower. However, the in-box version
of WSL cannot work with .vhdx files, and we run into trouble with
something (I am still not sure what) still using the ext4.vhdx files
for 1m+ after the distro is shut down when exporting. This does not
happen when exporting to .tar.

Signed-off-by: Adam Pickering <adam.pickering@suse.com>