Validate names before attempting to create the snapshot.
This lets us do more simplification with the snapshot
function-wrapper.

First, the wrapper no longer returns a func object.
You just call it, it calls the wrapped function as a
callback, and then it returns an error status | nil.

Second, we should be getting the app-paths once per
run.

Third, I have the manager.Create method verify that the
passed in name has been validated. There's a bit of overkill
here, but if we ever have a reason to create more than one
snapshot in a run, this code won't need changing.

Signed-off-by: Eric Promislow <epromislow@suse.com>