Simplify GetConnectionInfo() call
Remove all the special code to delay reporting when the
config file is missing because the user just executed
a factory reset; it doesn't seem to be used anywhere.

This commit also implements reading a `host` setting from
the config file, which seems to have been missing from the
original code.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>