WSL-Helper: Avoid clobbering Windows kubeconfig from Linux
In previous versions (Rancher Desktop up to 1.10), we created a symlink
from the Linux kubeconfig to the Windows kubeconfig.  We now instead use a
seprate file (as of RD 1.11); however, this can cause issues if the symlink
already existed, as we end up writing into the Windows file instead.

Fix that by explicitly removing any such symlinks (but not any unrelated
symlinks).

Signed-off-by: Mark Yen <mark.yen@suse.com>