WiX: Add executable for icon
WiX takes the executable to use for icons and makes a copy of it,
uncompressed, in the installer.  In our case that was a ~150MB executable
with the majority of Chromium.  Instead, create a stub executable with no
real code and embed the icon in that file.  This cuts the installer down
enough to fit in a CD.

Signed-off-by: Mark Yen <mark.yen@suse.com>