Don't cache connection settings from config file
When creating/restoring snapshots we may fetch the connection
settings multiple times, once to stop the backend, and once to
start it back up. The API credentials may change if the app is
restarted during that time, so cached values may no longer be
valid.

This changes makes sure that GetConnectionInfo() will update
settings from the file on each call, but still allow commandline
options to overwrite the file settings.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>