Do not add toString() to versions returned from AlpineLimaISO.getAvailableVersions()
When a newer version of the alpineLimaISO dependency is detected,
this version adheres to the AlpineLimaISOVersion type. However, in a
previous commit, a toString() method was added to the object returned
from AlpineLimaISO.getAvailableVersions(). toString() was not added
to the alpineLimaISO version that is read from dependencies.yaml.

Getting closer to what this commit is fixing: having a custom
toString() method defined on an object that we try to write to a file
using YAML.stringify() appears to cause the cryptic error:
"Error: Tag not resolved for Function value".

Putting it all together: only recently (2023-11-30) was a new version
of alpineLimaISO released. This is the first release of alpineLimaISO
since toString() was added to the versions returned from
AlpineLimaISO.getAvailableVersions() in
163cc1e91a38d3d555f7f8be632a87e70717e4c6.
Having toString() present only causes a problem when we try to write
the new alpineLimaISO to dependencies.yaml, which only happens when
we have found a new version of it and are creating a PR to bump it.
This is why the error only showed up now.

Signed-off-by: Adam Pickering <adam.pickering@suse.com>