Drop use of ctrlc-windows
The ctrlc-windows npm package contains a rust executable; rust by default
depends on vcruntime140.dll.  This means that on machines without the
correct MSVC redistributables installed, Rancher Desktop will fail to start
due to failing to import the native module.

Luckily, what we need from it is simple enough that we can just
re-implement the necessary functionality as part of wsl-helper.

Signed-off-by: Mark Yen <mark.yen@suse.com>