background: Ignore attempts to start from invalid states
If a snapshot is in progress when we start Rancher Desktop, we wait for the
snapshot progress to finish before continuing with starting.  This can lead
to an unexpected situation where we end up trying to start twice:

- Rancher Desktop (upon seeing the snapshot lock being freed) starts the
  backend.
- rdctl (upon finishing) fetches new API credentials and manually starts
  the backend.

This leads to errors as two independent flows try to start, causing issues
like terminating the WSL distribution halfway through starting.

This fix is done here to ensure that if the user manually starts the
backend via the API (or does so at points we do not expect) we check the
current state and prevent the start request if we are busy or already
started.

Signed-off-by: Mark Yen <mark.yen@suse.com>